==> notebooks/01_data_exploration.ipynb <==
import numpy as np
import pandas as pd

job_skills = pd.read_csv('C:/Users/ashua/Desktop/Inelligent Job Recomendation Engine/data/Jobs/job_skills.csv')
job_skills.head()

postings = pd.read_csv('C:/Users/ashua/Desktop/Inelligent Job Recomendation Engine/data/Jobs/postings.csv')
postings.head()

resume = pd.read_csv('C:/Users/ashua/Desktop/Inelligent Job Recomendation Engine/data/Resumes/resume.csv')
resume.head()

job_skills.info()

job_skills.isnull().sum()

job_skills.nunique()

job_skills.drop_duplicates(inplace=True)

job_skills.duplicated().sum()

job_skills.dropna(inplace=True)

unique_jobs = job_skills['job_id'].nunique()
print(f"Unique job IDs: {unique_jobs}")

postings.info()

postings.drop(['views', 'applies', 'original_listed_time', 'expiry', 'closed_time', 
                   'job_posting_url', 'application_url', 'application_type', 'pay_period', 
                   'max_salary', 'min_salary', 'med_salary', 'company_id', 'remote_allowed',
                   'listed_time', 'posting_domain', 'sponsored', 'currency', 
                   'compensation_type', 'normalized_salary', 'zip_code', 'fips',], axis=1, inplace=True)

postings.head()

postings.isnull().sum()

postings = postings.drop(columns=['skills_desc'])

postings = postings.dropna(subset=['description', 'company_name'])

postings['formatted_experience_level'] = postings['formatted_experience_level'].fillna('Not Specified')

postings.isnull().sum()

postings.duplicated().sum()

postings.nunique()

postings['description'] = postings['description'].str.lower().str.strip()

postings.info()

#Compamnies Represented 
company_counts = postings['company_name'].value_counts().head(10)
print(company_counts)

resume.info()

resume.isnull().sum()

resume.duplicated().sum()

resume['Resume_str'] = resume['Resume_str'].str.lower().str.strip()

# Cacluate word count in each resume
resume['word_count'] = resume['Resume_str'].apply(lambda x: len(x.split()))

#Average word count
average_word_count = resume['word_count'].mean()
print(f'Average Resume length: {average_word_count} words')

# minimum word count and maximum word count
min_word_count = resume['word_count'].min()
max_word_count = resume['word_count'].max()
print(f'Shortest Resume: {min_word_count} words')
print(f'Longest Resume: {max_word_count} words')

#Most Common Categories
common_categories = resume['Category'].value_counts()
print("Most Common Resume Category:")

print(f"\nMost common category: {common_categories.index[0]}")
print(f"Number of resumes: {common_categories.values[0]}")

#full distribution
category_dist = resume['Category'].value_counts()

# Show as table
print(category_dist)

resume.info()

postings.to_csv('C:/Users/ashua/Desktop/Inelligent Job Recomendation Engine/data/Jobs/postings_cleaned.csv', index=False)
job_skills.to_csv('C:/Users/ashua/Desktop/Inelligent Job Recomendation Engine/data/Jobs/job_skills_cleaned.csv', index=False)
resume.to_csv('C:/Users/ashua/Desktop/Inelligent Job Recomendation Engine/data/resumes/resume_cleaned.csv', index=False)

print("Cleaned data saved!")
